+++ src/types.ts
@@ -3,7 +3,7 @@
  */
 export type DecodeResult = {
 	sampleRate: number;
-	channelData: Float32Array[];
+	channelData: Float64Array[];
 };
 
 /**
